chore: refactor copy to make more readable (#12025)
* chore: refactor copy to make more readable

* rename to copy_commit_data_and_set_finish

* move table.rs from interpreter/common to pipeline/common

* decouple pipeline and interperteer

* better para order

* better func name

* common -> builders
