+++ src/query/service/src/interpreters/interpreter_copy.rs
@@ -12,27 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::collections::BTreeMap;
 use std::sync::Arc;
 use std::time::Instant;
 
-use chrono::Utc;
-use common_base::runtime::GlobalIORuntime;
 use common_catalog::plan::StageTableInfo;
 use common_catalog::table::AppendMode;
-use common_catalog::table::Table;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_expression::infer_table_schema;
 use common_expression::BlockThresholds;
 use common_expression::DataField;
-use common_expression::DataSchema;
 use common_expression::DataSchemaRef;
 use common_expression::DataSchemaRefExt;
-use common_expression::Scalar;
 use common_meta_app::principal::StageInfo;
-use common_meta_app::schema::TableCopiedFileInfo;
-use common_meta_app::schema::UpsertTableCopiedFileReq;
 use common_pipeline_core::Pipeline;
 use common_sql::executor::table_read_plan::ToReadDataSourcePlan;
 use common_sql::executor::DistributedCopyIntoTable;
@@ -43,19 +35,17 @@ use common_sql::plans::CopyIntoTableMode;
 use common_sql::plans::CopyIntoTablePlan;
 use common_storage::StageFileInfo;
 use common_storage::StageFilesInfo;
-use common_storages_fuse::io::Files;
 use common_storages_stage::StageTable;
-use tracing::debug;
-use tracing::error;
 use tracing::info;
 
-use crate::interpreters::common::append2table;
-use crate::interpreters::common::append2table_without_commit;
 use crate::interpreters::common::check_deduplicate_label;
 use crate::interpreters::Interpreter;
 use crate::interpreters::SelectInterpreter;
-use crate::pipelines::processors::transforms::TransformAddConstColumns;
-use crate::pipelines::processors::TransformCastSchema;
+use crate::pipelines::builders::build_append2table_pipeline;
+use crate::pipelines::builders::build_append_data_with_finish_pipeline;
+use crate::pipelines::builders::build_upsert_copied_files_to_meta_req;
+use crate::pipelines::builders::try_purge_files;
+use crate::pipelines::builders::CopyPlanParam;
 use crate::pipelines::PipelineBuildResult;
 use crate::schedulers::build_distributed_pipeline;
 use crate::sessions::QueryContext;
@@ -66,11 +56,6 @@ use crate::sql::plans::Plan;
 pub struct CopyInterpreter {
     ctx: Arc<QueryContext>,
     plan: CopyPlan,
-}
-
-pub enum PlanParam {
-    CopyIntoTablePlanOption(CopyIntoTablePlan),
-    DistributedCopyIntoTable(DistributedCopyIntoTable),
 }
 
 impl CopyInterpreter {
@@ -119,8 +104,9 @@ impl CopyInterpreter {
         Ok((build_res, data_schema))
     }
 
+    /// Build a pipeline for local copy into stage.
     #[async_backtrace::framed]
-    async fn build_copy_into_stage_pipeline(
+    async fn build_local_copy_into_stage_pipeline(
         &self,
         stage: &StageInfo,
         path: &str,
@@ -140,11 +126,11 @@ impl CopyInterpreter {
             is_select: false,
         };
         let table = StageTable::try_create(stage_table_info)?;
-        append2table(
+        build_append2table_pipeline(
             self.ctx.clone(),
+            &mut build_res.main_pipeline,
             table,
             data_schema,
-            &mut build_res.main_pipeline,
             None,
             false,
             AppendMode::Normal,
@@ -168,23 +154,8 @@ impl CopyInterpreter {
             info!(status);
         }
 
-        let table_ctx: Arc<dyn TableContext> = ctx.clone();
-        let op = StageTable::get_op(stage_info);
-        match op {
-            Ok(op) => {
-                let file_op = Files::create(table_ctx, op);
-                let files = stage_file_infos
-                    .iter()
-                    .map(|v| v.path.clone())
-                    .collect::<Vec<_>>();
-                if let Err(e) = file_op.remove_file_in_batch(&files).await {
-                    error!("Failed to delete file: {:?}, error: {}", files, e);
-                }
-            }
-            Err(e) => {
-                error!("Failed to get stage table op, error: {}", e);
-            }
-        }
+        try_purge_files(ctx.clone(), stage_info, stage_file_infos).await;
+
         // Status.
         info!(
             "end to purge files:{}, elapsed:{}",
@@ -199,7 +170,7 @@ impl CopyInterpreter {
     }
 
     #[async_backtrace::framed]
-    async fn transform_copy_plan_distributed(
+    async fn try_transform_copy_plan_from_local_to_distributed(
         &self,
         plan: &CopyIntoTablePlan,
     ) -> Result<Option<DistributedCopyIntoTable>> {
@@ -253,7 +224,7 @@ impl CopyInterpreter {
     }
 
     #[async_backtrace::framed]
-    async fn build_read_stage(
+    async fn build_read_stage_table_data_pipeline(
         &self,
         pipeline: &mut Pipeline,
         plan: &CopyIntoTablePlan,
@@ -290,8 +261,9 @@ impl CopyInterpreter {
         Ok(())
     }
 
+    /// Build a pipeline to copy data from local.
     #[async_backtrace::framed]
-    async fn build_copy_into_table_pipeline(
+    async fn build_local_copy_into_table_pipeline(
         &self,
         plan: &CopyIntoTablePlan,
     ) -> Result<PipelineBuildResult> {
@@ -318,7 +290,7 @@ impl CopyInterpreter {
             if files.is_empty() {
                 return Ok(build_res);
             }
-            self.build_read_stage(
+            self.build_read_stage_table_data_pipeline(
                 &mut build_res.main_pipeline,
                 plan,
                 to_table.get_block_thresholds(),
@@ -328,11 +300,11 @@ impl CopyInterpreter {
             (build_res, plan.required_source_schema.clone(), files)
         };
 
-        append_data_and_set_finish(
+        build_append_data_with_finish_pipeline(
+            ctx,
             &mut build_res.main_pipeline,
             source_schema,
-            PlanParam::CopyIntoTablePlanOption(plan.clone()),
-            ctx,
+            CopyPlanParam::CopyIntoTablePlanOption(plan.clone()),
             to_table,
             files,
             start,
@@ -341,53 +313,46 @@ impl CopyInterpreter {
         Ok(build_res)
     }
 
-    fn upsert_copied_files_request(
-        ctx: Arc<QueryContext>,
-        to_table: Arc<dyn Table>,
-        stage_info: StageInfo,
-        copied_files: Vec<StageFileInfo>,
-        force: bool,
-    ) -> Result<Option<UpsertTableCopiedFileReq>> {
-        let mut copied_file_tree = BTreeMap::new();
-        for file in &copied_files {
-            // Short the etag to 7 bytes for less space in metasrv.
-            let short_etag = file.etag.clone().map(|mut v| {
-                v.truncate(7);
-                v
-            });
-            copied_file_tree.insert(file.path.clone(), TableCopiedFileInfo {
-                etag: short_etag,
-                content_length: file.size,
-                last_modified: Some(file.last_modified),
-            });
+    /// Build a pipeline to copy data into table for distributed.
+    #[async_backtrace::framed]
+    async fn build_distributed_copy_into_table_pipeline(
+        &self,
+        distributed_plan: &DistributedCopyIntoTable,
+    ) -> Result<PipelineBuildResult> {
+        // add exchange plan node to enable distributed
+        // TODO(leiysky): we reuse the id of exchange here,
+        // which is not correct. We should generate a new id for insert.
+        let exchange_plan = PhysicalPlan::Exchange(Exchange {
+            plan_id: 0,
+            input: Box::new(PhysicalPlan::DistributedCopyIntoTable(Box::new(
+                distributed_plan.clone(),
+            ))),
+            kind: FragmentKind::Merge,
+            keys: Vec::new(),
+        });
+        let mut build_res = build_distributed_pipeline(&self.ctx, &exchange_plan, false).await?;
+
+        let catalog = self.ctx.get_catalog(&distributed_plan.catalog_name)?;
+        let to_table = catalog.get_table_by_info(&distributed_plan.table_info)?;
+        let copied_files = build_upsert_copied_files_to_meta_req(
+            self.ctx.clone(),
+            to_table.clone(),
+            distributed_plan.stage_table_info.stage_info.clone(),
+            distributed_plan.files.clone(),
+            distributed_plan.force,
+        )?;
+        let mut overwrite_ = false;
+        if let CopyIntoTableMode::Insert { overwrite } = distributed_plan.write_mode {
+            overwrite_ = overwrite;
         }
+        to_table.commit_insertion(
+            self.ctx.clone(),
+            &mut build_res.main_pipeline,
+            copied_files,
+            overwrite_,
+        )?;
 
-        let expire_hours = ctx.get_settings().get_load_file_metadata_expire_hours()?;
-
-        let upsert_copied_files_request = {
-            if stage_info.copy_options.purge && force {
-                // if `purge-after-copy` is enabled, and in `force` copy mode,
-                // we do not need to upsert copied files into meta server
-                info!(
-                    "[purge] and [force] are both enabled,  will not update copied-files set. ({})",
-                    &to_table.get_table_info().desc
-                );
-                None
-            } else if copied_file_tree.is_empty() {
-                None
-            } else {
-                tracing::debug!("upsert_copied_files_info: {:?}", copied_file_tree);
-                let expire_at = expire_hours * 60 * 60 + Utc::now().timestamp() as u64;
-                let req = UpsertTableCopiedFileReq {
-                    file_info: copied_file_tree,
-                    expire_at: Some(expire_at),
-                    fail_if_duplicated: !force,
-                };
-                Some(req)
-            }
-        };
-
-        Ok(upsert_copied_files_request)
+        Ok(build_res)
     }
 }
 
@@ -407,263 +372,26 @@ impl Interpreter for CopyInterpreter {
         match &self.plan {
             CopyPlan::IntoTable(plan) => {
                 if plan.enable_distributed {
-                    let distributed_plan_op = self.transform_copy_plan_distributed(plan).await?;
-                    if distributed_plan_op.is_none() {
-                        return self.build_copy_into_table_pipeline(plan).await;
+                    let distributed_plan_op = self
+                        .try_transform_copy_plan_from_local_to_distributed(plan)
+                        .await?;
+                    if let Some(distributed_plan) = distributed_plan_op {
+                        self.build_distributed_copy_into_table_pipeline(&distributed_plan)
+                            .await
+                    } else {
+                        self.build_local_copy_into_table_pipeline(plan).await
                     }
-                    let distributed_plan = distributed_plan_op.unwrap();
-
-                    // add exchange plan node to enable distributed
-                    // TODO(leiysky): we reuse the id of exchange here,
-                    // which is not correct. We should generate a new id for insert.
-                    let exchange_plan = PhysicalPlan::Exchange(Exchange {
-                        plan_id: 0,
-                        input: Box::new(PhysicalPlan::DistributedCopyIntoTable(Box::new(
-                            distributed_plan.clone(),
-                        ))),
-                        kind: FragmentKind::Merge,
-                        keys: Vec::new(),
-                    });
-                    let mut build_res =
-                        build_distributed_pipeline(&self.ctx, &exchange_plan, false).await?;
-
-                    let catalog = self.ctx.get_catalog(&distributed_plan.catalog_name)?;
-                    let to_table = catalog.get_table_by_info(&distributed_plan.table_info)?;
-                    let copied_files = CopyInterpreter::upsert_copied_files_request(
-                        self.ctx.clone(),
-                        to_table.clone(),
-                        distributed_plan.stage_table_info.stage_info.clone(),
-                        distributed_plan.files.clone(),
-                        distributed_plan.force,
-                    )?;
-                    let mut overwrite_ = false;
-                    if let CopyIntoTableMode::Insert { overwrite } = plan.write_mode {
-                        overwrite_ = overwrite;
-                    }
-                    to_table.commit_insertion(
-                        self.ctx.clone(),
-                        &mut build_res.main_pipeline,
-                        copied_files,
-                        overwrite_,
-                    )?;
-
-                    Ok(build_res)
                 } else {
-                    self.build_copy_into_table_pipeline(plan).await
+                    self.build_local_copy_into_table_pipeline(plan).await
                 }
             }
             CopyPlan::IntoStage {
                 stage, from, path, ..
-            } => self.build_copy_into_stage_pipeline(stage, path, from).await,
+            } => {
+                self.build_local_copy_into_stage_pipeline(stage, path, from)
+                    .await
+            }
             CopyPlan::NoFileToCopy => Ok(PipelineBuildResult::create()),
         }
     }
-}
-
-fn fill_const_columns(
-    ctx: Arc<QueryContext>,
-    pipeline: &mut Pipeline,
-    input_schema: DataSchemaRef,
-    output_schema: DataSchemaRef,
-    const_values: Vec<Scalar>,
-) -> Result<()> {
-    pipeline.add_transform(|transform_input_port, transform_output_port| {
-        TransformAddConstColumns::try_create(
-            ctx.clone(),
-            transform_input_port,
-            transform_output_port,
-            input_schema.clone(),
-            output_schema.clone(),
-            const_values.clone(),
-        )
-    })?;
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn append_data_and_set_finish(
-    main_pipeline: &mut Pipeline,
-    source_schema: Arc<DataSchema>,
-    plan_option: PlanParam,
-    ctx: Arc<QueryContext>,
-    to_table: Arc<dyn Table>,
-    files: Vec<StageFileInfo>,
-    start: Instant,
-    use_commit: bool,
-) -> Result<()> {
-    let plan_required_source_schema: DataSchemaRef;
-    let plan_required_values_schema: DataSchemaRef;
-    let plan_values_consts: Vec<Scalar>;
-    let plan_stage_table_info: StageTableInfo;
-    let plan_force: bool;
-    let plan_write_mode: CopyIntoTableMode;
-    let local_id;
-    match plan_option {
-        PlanParam::CopyIntoTablePlanOption(plan) => {
-            plan_required_source_schema = plan.required_source_schema;
-            plan_required_values_schema = plan.required_values_schema;
-            plan_values_consts = plan.values_consts;
-            plan_stage_table_info = plan.stage_table_info;
-            plan_force = plan.force;
-            plan_write_mode = plan.write_mode;
-            local_id = ctx.get_cluster().local_id.clone();
-        }
-        PlanParam::DistributedCopyIntoTable(plan) => {
-            plan_required_source_schema = plan.required_source_schema;
-            plan_required_values_schema = plan.required_values_schema;
-            plan_values_consts = plan.values_consts;
-            plan_stage_table_info = plan.stage_table_info;
-            plan_force = plan.force;
-            plan_write_mode = plan.write_mode;
-            local_id = plan.local_node_id;
-        }
-    }
-
-    debug!("source schema:{:?}", source_schema);
-    debug!("required source schema:{:?}", plan_required_source_schema);
-    debug!("required values schema:{:?}", plan_required_values_schema);
-
-    if source_schema != plan_required_source_schema {
-        // only parquet need cast
-        let func_ctx = ctx.get_function_context()?;
-        main_pipeline.add_transform(|transform_input_port, transform_output_port| {
-            TransformCastSchema::try_create(
-                transform_input_port,
-                transform_output_port,
-                source_schema.clone(),
-                plan_required_source_schema.clone(),
-                func_ctx.clone(),
-            )
-        })?;
-    }
-
-    if !plan_values_consts.is_empty() {
-        fill_const_columns(
-            ctx.clone(),
-            main_pipeline,
-            source_schema,
-            plan_required_values_schema.clone(),
-            plan_values_consts,
-        )?;
-    }
-
-    let stage_info_clone = plan_stage_table_info.stage_info;
-    let write_mode = plan_write_mode;
-    let mut purge = true;
-    match write_mode {
-        CopyIntoTableMode::Insert { overwrite } => {
-            if use_commit {
-                append2table(
-                    ctx.clone(),
-                    to_table,
-                    plan_required_values_schema,
-                    main_pipeline,
-                    None,
-                    overwrite,
-                    AppendMode::Copy,
-                )?;
-            } else {
-                append2table_without_commit(
-                    ctx.clone(),
-                    to_table,
-                    plan_required_values_schema,
-                    main_pipeline,
-                    AppendMode::Copy,
-                )?
-            }
-        }
-        CopyIntoTableMode::Copy => {
-            if !stage_info_clone.copy_options.purge {
-                purge = false;
-            }
-
-            if use_commit {
-                let copied_files = CopyInterpreter::upsert_copied_files_request(
-                    ctx.clone(),
-                    to_table.clone(),
-                    stage_info_clone.clone(),
-                    files.clone(),
-                    plan_force,
-                )?;
-                append2table(
-                    ctx.clone(),
-                    to_table,
-                    plan_required_values_schema,
-                    main_pipeline,
-                    copied_files,
-                    false,
-                    AppendMode::Copy,
-                )?;
-            } else {
-                append2table_without_commit(
-                    ctx.clone(),
-                    to_table,
-                    plan_required_values_schema,
-                    main_pipeline,
-                    AppendMode::Copy,
-                )?
-            }
-        }
-        CopyIntoTableMode::Replace => {}
-    }
-    if local_id == ctx.get_cluster().local_id {
-        main_pipeline.set_on_finished(move |may_error| {
-            match may_error {
-                None => {
-                    GlobalIORuntime::instance().block_on(async move {
-                        {
-                            let status =
-                                format!("end of commit, number of copied files:{}", files.len());
-                            ctx.set_status_info(&status);
-                            info!(status);
-                        }
-
-                        // 1. log on_error mode errors.
-                        // todo(ariesdevil): persist errors with query_id
-                        if let Some(error_map) = ctx.get_maximum_error_per_file() {
-                            for (file_name, e) in error_map {
-                                error!(
-                                    "copy(on_error={}): file {} encounter error {},",
-                                    stage_info_clone.copy_options.on_error,
-                                    file_name,
-                                    e.to_string()
-                                );
-                            }
-                        }
-
-                        // 2. Try to purge copied files if purge option is true, if error will skip.
-                        // If a file is already copied(status with AlreadyCopied) we will try to purge them.
-                        if purge {
-                            CopyInterpreter::try_purge_files(
-                                ctx.clone(),
-                                &stage_info_clone,
-                                &files,
-                            )
-                            .await;
-                        }
-
-                        // Status.
-                        {
-                            info!("all copy finished, elapsed:{}", start.elapsed().as_secs());
-                        }
-
-                        Ok(())
-                    })?;
-                }
-                Some(error) => {
-                    error!(
-                        "copy failed, elapsed:{}, reason: {}",
-                        start.elapsed().as_secs(),
-                        error
-                    );
-                }
-            }
-            Ok(())
-        });
-    } else {
-        // remote node does nothing.
-        main_pipeline.set_on_finished(move |_| Ok(()))
-    }
-
-    Ok(())
 }
