+++ src/query/service/src/pipelines/pipeline_builder.rs
@@ -91,9 +91,9 @@ use super::processors::transforms::WindowFunctionInfo;
 use super::processors::TransformExpandGroupingSets;
 use crate::api::DefaultExchangeInjector;
 use crate::api::ExchangeInjector;
-use crate::interpreters::append_data_and_set_finish;
-use crate::interpreters::fill_missing_columns;
-use crate::interpreters::PlanParam;
+use crate::pipelines::builders::build_append_data_with_finish_pipeline;
+use crate::pipelines::builders::build_fill_missing_columns_pipeline;
+use crate::pipelines::builders::CopyPlanParam;
 use crate::pipelines::processors::transforms::build_partition_bucket;
 use crate::pipelines::processors::transforms::AggregateInjector;
 use crate::pipelines::processors::transforms::FinalSingleStateAggregator;
@@ -225,11 +225,11 @@ impl PipelineBuilder {
         let start = Instant::now();
         stage_table.read_data(table_ctx, &distributed_plan.source, &mut self.main_pipeline)?;
         // append data
-        append_data_and_set_finish(
+        build_append_data_with_finish_pipeline(
+            ctx,
             &mut self.main_pipeline,
             distributed_plan.required_source_schema.clone(),
-            PlanParam::DistributedCopyIntoTable(distributed_plan.clone()),
-            ctx,
+            CopyPlanParam::DistributedCopyIntoTable(distributed_plan.clone()),
             to_table,
             distributed_plan.files.clone(),
             start,
@@ -1384,11 +1384,11 @@ impl PipelineBuilder {
             .get_table_by_info(&insert_select.table_info)?;
 
         let source_schema = insert_schema;
-        fill_missing_columns(
+        build_fill_missing_columns_pipeline(
             self.ctx.clone(),
+            &mut self.main_pipeline,
             table.clone(),
             source_schema.clone(),
-            &mut self.main_pipeline,
         )?;
 
         table.append_data(
